resolved make run issue with module
