+++ src/api/resources/airport.py
@@ -1,9 +1,8 @@
 from pydantic import BaseModel
-from typing import List
 
-from api.resources.runway import Runway
+from src.api.resources.runway import Runway
 class Airport(BaseModel):
     altitude: int
-    runways: List[Runway]
+    runways: list[Runway]
     total_arrivals: int
     total_departures: int
